Moves our HTTP things and separates the httpx from the rest
