+++ fauna/__init__.py
@@ -7,7 +7,7 @@ __copyright__ = "2023 Fauna, Inc"
 
 from .client import Client
 from .headers import Header
-from .http_client import HTTPClient, HTTPResponse, HTTPXClient
+from .http import HTTPClient, HTTPResponse, HTTPXClient
 from .query_builder import fql, QueryInterpolation
 from .models import Document, DocumentReference, NamedDocument, NamedDocumentReference, Module
 from .errors import AuthenticationError, ClientError, ProtocolError, ServiceError, AuthorizationError, \
