+++ fauna/http/__init__.py
@@ -1,0 +1,2 @@
+from .http_client import HTTPClient, HTTPResponse
+from .httpx_client import HTTPXClient
