+++ fauna/http/httpx_client.py
@@ -1,76 +1,10 @@
-import abc
 import json
-from typing import Optional, Iterator, Mapping, Any
-from dataclasses import dataclass
+from typing import Mapping, Any, Optional, Iterator
 
 import httpx
 
-from .errors import ClientError, NetworkError
-
-
-@dataclass(frozen=True)
-class ErrorResponse:
-    status_code: int
-    error_code: str
-    error_message: str
-    summary: str
-
-
-class HTTPResponse(abc.ABC):
-
-    @abc.abstractmethod
-    def headers(self) -> Mapping[str, str]:
-        pass
-
-    @abc.abstractmethod
-    def status_code(self) -> int:
-        pass
-
-    @abc.abstractmethod
-    def json(self) -> Any:
-        pass
-
-    @abc.abstractmethod
-    def read(self) -> bytes:
-        pass
-
-    @abc.abstractmethod
-    def iter_bytes(self) -> Iterator[bytes]:
-        pass
-
-    @abc.abstractmethod
-    def close(self):
-        pass
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.close()
-
-
-class HTTPClient(abc.ABC):
-
-    # TODO(lucas): initialize base class with default retry strategy
-
-    @abc.abstractmethod
-    def request(
-        self,
-        method: str,
-        url: str,
-        headers: Mapping[str, str],
-        data: Mapping[str, Any],
-    ) -> HTTPResponse:
-        pass
-
-    @abc.abstractmethod
-    def stream(
-        self,
-        url: str,
-        headers: Mapping[str, str],
-        data: Mapping[str, Any],
-    ) -> HTTPResponse:
-        pass
+from fauna.errors import ClientError, NetworkError
+from fauna.http.http_client import HTTPResponse, HTTPClient
 
 
 class HTTPXResponse(HTTPResponse):
